+++ src/Makefile
@@ -1,15 +1,14 @@
 
 SRC_OBJS	= virtual.o format.o period.o player.o read_event.o dataio.o \
-		  misc.o mkstemp.o xfnmatch.o md5.o lfo.o scan.o control.o \
+		  misc.o mkstemp.o md5.o lfo.o scan.o control.o \
 		  med_extras.o filter.o effects.o mixer.o mix_all.o \
 		  load_helpers.o load.o hio.o hmn_extras.o extras.o smix.o \
 		  memio.o tempfile.o mix_paula.o win32.o
 
 SRC_DFILES	= Makefile $(SRC_OBJS:.o=.c) common.h effects.h \
 		  format.h lfo.h list.h mixer.h period.h player.h virtual.h \
-		  xfnmatch.h md5.h precomp_lut.h tempfile.h med_extras.h hio.h \
-		  hmn_extras.h extras.h memio.h mdataio.h depacker.h paula.h \
-		  precomp_blep.h
+		  md5.h precomp_lut.h tempfile.h med_extras.h hio.h \
+		  hmn_extras.h extras.h memio.h mdataio.h paula.h precomp_blep.h
 
 SRC_PATH	= src
