+++ src/depackers/Makefile
@@ -1,13 +1,13 @@
 
-DEPACKERS_OBJS	= ppdepack.o unsqsh.o mmcmp.o readrle.o readlzw.o \
+DEPACKERS_OBJS	= depacker.o ppdepack.o unsqsh.o mmcmp.o readrle.o readlzw.o \
 		  unarc.o arcfs.o xfd.o inflate.o muse.o unlzx.o s404_dec.o \
 		  unzip.o gunzip.o uncompress.o unxz.o bunzip2.o unlha.o \
 		  xz_dec_lzma2.o xz_dec_stream.o oxm.o vorbis.o crc32.o \
-		  xfd_link.o
+		  xfd_link.o xfnmatch.o
 
-DEPACKERS_DFILES = Makefile $(DEPACKERS_OBJS:.o=.c) readhuff.h readlzw.h \
+DEPACKERS_DFILES = Makefile $(DEPACKERS_OBJS:.o=.c) depacker.h readhuff.h readlzw.h \
 		  readrle.h inflate.h xz_lzma2.h README.unxz xz.h \
-		  xz_private.h xz_stream.h xz_config.h vorbis.h crc32.h
+		  xz_private.h xz_stream.h xz_config.h vorbis.h crc32.h xfnmatch.h
 
 DEPACKERS_PATH	= src/depackers
