+++ src/depackers/depacker.h
@@ -1,7 +1,8 @@
-#ifndef XMP_DEPACKER_H
-#define XMP_DEPACKER_H
+#ifndef LIBXMP_DEPACKER_H
+#define LIBXMP_DEPACKER_H
 
 #include "common.h"
+#include "hio.h"
 
 extern struct depacker libxmp_depacker_zip;
 extern struct depacker libxmp_depacker_lha;
@@ -25,4 +26,7 @@ struct depacker {
 	int (*const depack)(FILE *, FILE *);
 };
 
-#endif
+int	libxmp_decrunch		(HIO_HANDLE **h, const char *filename, char **temp);
+int	libxmp_exclude_match	(const char *);
+
+#endif /* LIBXMP_DEPACKER_H */
