+++ test-dev/Makefile.in
@@ -583,7 +583,7 @@ TEST_PATH	= .
 SRC_PATH	= ../src
 
 TEST_INTERNAL	= md5.o win32.o hio.o load_helpers.o loaders/itsex.o dataio.o scan.o \
-		  loaders/sample.o loaders/common.o period.o xfnmatch.o memio.o
+		  loaders/sample.o loaders/common.o period.o depackers/xfnmatch.o memio.o
 
 T_OBJS 		= $(addprefix $(TEST_PATH)/,$(TEST_OBJS)) \
 		  $(addprefix $(SRC_PATH)/,$(TEST_INTERNAL))
